% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratMarkerFinder.R
\name{addDAslot}
\alias{addDAslot}
\title{Add DA slot}
\usage{
addDAslot(object, da.regions, da.slot = "da", set.ident = F)
}
\arguments{
\item{object}{input Seurat object}

\item{da.regions}{output from function getDAregion()}

\item{da.slot}{character, variable name to put in Seurat meta.data, default "da"}

\item{set.ident}{a logical value to indicate whether to set Idents of the Seurat object to DA information,
default False}
}
\value{
updated Seurat object
}
\description{
Add DA region information to the meta.data of a Seurat object
}
