% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAseq.R
\docType{package}
\name{DAseq}
\alias{DAseq}
\alias{DAseq-package}
\title{DAseq: Detecting regions of differential abundance between scRNA-seq datasets}
\description{
DA-seq is a multiscale approach for detecting DA subpopulations.
In contrast to clustering based approaches, our method can detect DA subpopulations that do not form
well separated clusters.
For each cell, we compute a multiscale differential abundance score measure.
These scores are based on the k nearest neighbors in the transcriptome space for various values of k.
}
\author{
Jun Zhao, Ariel Jaffe, Henry Li, Ofir Lindenbaum, Xiuyuan Cheng, Yuval Kluger
}
