% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDAcells.R
\name{getDAcells}
\alias{getDAcells}
\title{DA-seq Step 1 & Step 2: select DA cells}
\usage{
getDAcells(X, cell.labels, labels.1, labels.2, k.vector = NULL,
  save.knn = F, alpha = 0, k.folds = 10, n.runs = 5, n.rand = 2,
  pred.thres = NULL, do.plot = T, plot.embedding = NULL,
  size = 0.5)
}
\arguments{
\item{X}{size N-by-p matrix, input merged dataset of interest after dimension reduction.}

\item{cell.labels}{size N character vector, labels for each input cell}

\item{labels.1}{character vector, label name(s) that represent condition 1}

\item{labels.2}{character vector, label name(s) that represent condition 2}

\item{k.vector}{vector, k values to create the score vector}

\item{save.knn}{a logical value to indicate whether to save computed kNN result, default False}

\item{alpha}{numeric, elasticnet mixing parameter passed to glmnet(), default 0 (Ridge)}

\item{k.folds}{integer, number of data splits used in the neural network, default 10}

\item{n.runs}{integer, number of times to run the neural network to get the predictions, default 5}

\item{n.rand}{integer, number of random permutations to run, default 2}

\item{pred.thres}{length-2 vector, top and bottom threshold on DA measure,
default NULL, select significant DA cells based on permutation}

\item{do.plot}{a logical value to indicate whether to return ggplot objects showing the results,
default True}

\item{plot.embedding}{size N-by-2 matrix, 2D embedding for the cells}

\item{size}{cell size to use in the plot, default 0.5}
}
\value{
a list of results
\describe{
  \item{cell.idx}{index of cells used in DA calculation}
  \item{da.ratio}{score vector for each cell}
  \item{da.pred}{(mean) prediction from the logistic regression}
  \item{da.up}{index for DA cells more abundant in condition of labels.2}
  \item{da.down}{index for DA cells more abundant in condition of labels.1}
  \item{pred.plot}{ggplot object showing the predictions of logistic regression on plot.embedding}
  \item{da.cells.plot}{ggplot object highlighting cells of da.cell.idx on plot.embedding}
}
}
\description{
Step 1: compute a multiscale score measure for each cell of its k-nearest-neighborhood for
multiple values of k.
Step 2: train a logistic regression classifier based on the multiscale score measure and retain cells
that may reside in DA regions.
}
