% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDAsite}
\alias{plotDAsite}
\title{Plot da site}
\usage{
plotDAsite(X, site.list, size = 0.5, cols = NULL)
}
\arguments{
\item{X}{matrix, 2D embedding of each cell for the plot}

\item{site.list}{list, a list of cell indices for each site to plot}

\item{size}{numeric, dot size for each cell, default 0.5}

\item{cols}{string vector, color bar to use for each site, default ggplot default}
}
\description{
Plot da site
}
