% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STGmarkerFinder.R
\name{runSTG}
\alias{runSTG}
\title{Run STG}
\usage{
runSTG(X, X.labels, lambda = 1.5, n.runs = 5, return.model = F,
  python.use = "/usr/bin/python", GPU = "")
}
\arguments{
\item{X}{matrix, normalized expression matrix of all cells in the dataset, genes are in rows,
rownames must be gene names}

\item{X.labels}{numeric vector, specify labels for each cell, must be 0 or 1}

\item{lambda}{numeric, regularization parameter that weights the number of selected genes,
a larger lambda leads to fewer genes, default 1.5}

\item{n.runs}{integer, number of runs to run the model, default 5}

\item{return.model}{a logical value to indicate whether to return the actual model of STG}

\item{python.use}{character string, the Python to use, default "/usr/bin/python"}

\item{GPU}{which GPU to use, default '', using CPU}
}
\value{
a list of results:
\describe{
  \item{markers}{a list of data.frame with markers for each DA region}
  \item{accuracy}{a numeric vector showing mean accuracy for each DA region}
  \item{model}{a list of model for each DA region, each model contains:
    \describe{\item{model}{the model of STG of the final run}
    \item{features}{features used to train the model}
    \item{selected.features}{the selected features of the final run}
    \item{pred}{the linear prediction value for each cell from the model}}
  }
}
}
\description{
Run STG to select a set of genes that separate cells with one label from the other label
}
