% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratMarkerFinder.R
\name{SeuratLocalMarkers}
\alias{SeuratLocalMarkers}
\title{Find local markers}
\usage{
SeuratLocalMarkers(object, da.slot = "da", da.region.to.run,
  cell.label.slot, cell.label.to.run, ...)
}
\arguments{
\item{object}{input Seurat object}

\item{da.slot}{character, variable name that represents DA regions in Seurat meta.data, default "da"}

\item{da.region.to.run}{numeric, which (single) DA region to find local markers for}

\item{cell.label.slot}{character, variable name that represents cell labeling information in Seurat meta.data
to combine with DA information}

\item{cell.label.to.run}{cell label(s) that represent the local neiborhood for the input DA region}

\item{...}{parameters passed to Seurat FindMarkers() function}
}
\value{
a data.frame of markers and statistics
}
\description{
Use Seurat FindMarkers() function to identify genes that distinguish a DA region from its local neighborhood
}
