% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STGmarkerFinder.R
\name{STGlocalMarkers}
\alias{STGlocalMarkers}
\title{STG local markers
Run STG to find a set of genes that separate a given DA region from a local subset of cells.}
\usage{
STGlocalMarkers(X, da.regions, da.region.to.run, cell.label.info,
  cell.label.to.run, ...)
}
\arguments{
\item{X}{matrix, normalized expression matrix of all cells in the dataset, genes are in rows,
rownames must be gene names}

\item{da.regions}{output from the function getDAregion()}

\item{da.region.to.run}{numeric, which (single) DA region to find local markers for}

\item{cell.label.info}{vector, cell labeling information to select the local subset of cells to compare
with input DA region}

\item{cell.label.to.run}{cell label(s) to select from cell.label.info that represent
the local neiborhood for the input DA region}

\item{lambda}{numeric, regularization parameter that weights the number of selected genes,
a larger lambda leads to fewer genes, default 1.5}

\item{n.runs}{integer, number of runs to run the model, default 5}

\item{python.use}{character string, the Python to use, default "/usr/bin/python"}

\item{return.model}{a logical value to indicate whether to return the actual model of STG}

\item{GPU}{which GPU to use, default '', using CPU}
}
\value{
a list of results:
\describe{
  \item{markers}{a list of data.frame with markers for each DA region}
  \item{accuracy}{a numeric vector showing mean accuracy for each DA region}
  \item{model}{a list of model for each DA region, each model contains:
    \describe{\item{model}{the model of STG of the final run}
    \item{features}{features used to train the model}
    \item{selected.features}{the selected features of the final run}
    \item{pred}{the linear prediction value for each cell from the model}}
  }
}
}
\description{
STG local markers
Run STG to find a set of genes that separate a given DA region from a local subset of cells.
}
