% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STGmarkerFinder.R
\name{STGmarkerFinder}
\alias{STGmarkerFinder}
\title{DA-seq Step 4: STG feature selection}
\usage{
STGmarkerFinder(X, cell.idx, da.region.label, da.regions.to.run = NULL,
  lambda = 1.2, n.runs = 5, return.model = F,
  python.use = "/usr/bin/python", GPU = "")
}
\arguments{
\item{X}{matrix, normalized expression matrix of all cells in the dataset, genes are in rows,
rownames must be gene names}

\item{cell.idx}{result "da.cell.idx" from the output of function getDAcells}

\item{da.region.label}{result "cluster.res" from the output of function getDAregion}

\item{da.regions.to.run}{numeric (vector), which DA regions to run the marker finder,
default is to run all regions}

\item{lambda}{numeric, regularization parameter that weights the number of selected genes,
a larger lambda leads to fewer genes, default 1.2}

\item{n.runs}{integer, number of runs to run the model, default 5}

\item{return.model}{a logical value to indicate whether to return the actual model of STG}

\item{python.use}{character string, the Python to use, default "/usr/bin/python"}

\item{GPU}{which GPU to use, default '', using CPU}
}
\value{
a list of results:
\describe{
  \item{da.markers}{a list of data.frame with markers for each DA region}
  \item{accuracy}{a numeric vector showing mean accuracy for each DA region}
  \item{model}{a list of model for each DA region, each model contains:
    \describe{\item{model}{the model of STG of the final run}
    \item{features}{features used to train the model}
    \item{selected.features}{the selected features of the final run}
    \item{pred}{the linear prediction value for each cell from the model}}
  }
}
}
\description{
Use STG (stochastic gates) to select genes that separate each DA region from the rest of the cells.
For a full description of the algorithm, see Y. Yamada, O. Lindenbaum, S. Negahban, and Y. Kluger.
Feature selection using stochastic gates. arXiv preprint arXiv:1810.04247, 2018.
}
