% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X.2d.melanoma}
\alias{X.2d.melanoma}
\title{t-SNE embedding of the melanoma dataset}
\format{An object of class \code{matrix} with 16291 rows and 2 columns.}
\source{
\href{https://www.sciencedirect.com/science/article/pii/S0092867418313941}
{Sade-Feldman, Moshe, et al. (Cell. 2018)}
}
\usage{
X.2d.melanoma
}
\description{
A matrix containing 2D t-SNE embedding of the melanoma dataset
}
\keyword{datasets}
