% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X.label.info}
\alias{X.label.info}
\title{Sample label information}
\format{a data.frame with 48 rows and 2 columns
\describe{
  \item{label}{sample label, matching with labels in X.label.melanoma}
  \item{condition}{condition of the sample label, either R or NR}
}}
\source{
\href{https://www.sciencedirect.com/science/article/pii/S0092867418313941}
{Sade-Feldman, Moshe, et al. (Cell. 2018)}
}
\usage{
X.label.info
}
\description{
A dataset containing information of each sample, indicating whether this sample is a "responder" (R) or
a "non-responder" (NR)
}
\keyword{datasets}
