% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X.label.melanoma}
\alias{X.label.melanoma}
\title{Cell sample labels of the melanoma dataset}
\format{An object of class \code{character} of length 16291.}
\source{
\href{https://www.sciencedirect.com/science/article/pii/S0092867418313941}
{Sade-Feldman, Moshe, et al. (Cell. 2018)}
}
\usage{
X.label.melanoma
}
\description{
A string vector with the length equal to number of cells, indicating sample labels of each cell: which
sample each cell comes from
}
\keyword{datasets}
