% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X.melanoma}
\alias{X.melanoma}
\title{Top 10 PCs of the melanoma dataset}
\format{An object of class \code{matrix} with 16291 rows and 10 columns.}
\source{
\href{https://www.sciencedirect.com/science/article/pii/S0092867418313941}
{Sade-Feldman, Moshe, et al. (Cell. 2018)}
}
\usage{
X.melanoma
}
\description{
A dataset containing the top 10 PCs (principal components) of the melanoma dataset
}
\keyword{datasets}
