% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDAregion.R
\name{getDAregion}
\alias{getDAregion}
\title{DA-seq Step 3: get DA regions}
\usage{
getDAregion(X, cell.idx, k, alpha, restr.fact = 50, cell.labels,
  labels.1, labels.2, do.plot = T, plot.embedding = NULL, size = 0.5,
  seed = 0, ...)
}
\arguments{
\item{X}{size N-by-p matrix, input merged dataset of interest after dimension reduction}

\item{cell.idx}{result "da.cell.idx" from the output of function getDAcells}

\item{k}{number of DA regions to find for cells from function getDAcells}

\item{alpha}{estimated ratio of outliers of cells from function getDAcells}

\item{restr.fact}{parameter inherited from function "tclust"}

\item{cell.labels}{size N vector, labels for each input cell}

\item{labels.1}{vector, label name(s) that represent condition 1}

\item{labels.2}{vector, label name(s) that represent condition 2}

\item{do.plot}{a logical value to indicate whether to return ggplot objects showing the results, default True}

\item{plot.embedding}{size N-by-2 matrix, 2D embedding for the cells}

\item{size}{cell size to use in the plot, default 0.5}

\item{...}{other parameters passed to function tclust}
}
\value{
a list of results
\describe{
  \item{cluster.res}{DA region number for each cell from cell.idx, '0' represents outlier cells}
  \item{DA.stat}{a table showing DA score and p-value for each DA region}
  \item{da.region.plot}{ggplot object showing DA regions on plot.embedding}
}
}
\description{
Cluster the DA cells retained from Step 1 and Step 2 of DA-seq with outlier removal to obtain spatially
coherent DA regions.
}
