% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCellLabel}
\alias{plotCellLabel}
\title{Plot cell labels}
\usage{
plotCellLabel(X, label, cell.col = NULL, size = 0.5, do.label = T)
}
\arguments{
\item{X}{matrix, 2D embedding of each cell for the plot}

\item{label}{vector, label for each cell}

\item{cell.col}{string vector, color bar to use for cell labels, default ggplot default}

\item{size}{numeric, dot size for each cell, default 0.5}

\item{do.label}{a logical value indicating whether to add text to mark each cell label}
}
\value{
a ggplot object
}
\description{
Produce a ggplot object with cells on 2D embedding, colored by given labels of each cell.
}
