% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCellLabel}
\alias{plotCellLabel}
\title{Plot cell labels}
\usage{
plotCellLabel(X, label, cell.col = NULL, size = 0.5, alpha = 1,
  shape = 16, do.label = T, label.size = 4, label.repel = F,
  label.plot = NULL)
}
\arguments{
\item{X}{matrix, 2D embedding of each cell for the plot}

\item{label}{vector, label for each cell}

\item{cell.col}{string vector, color bar to use for cell labels, default ggplot default}

\item{size}{numeric, dot size for each cell, default 0.5}

\item{alpha}{numeric between 0 to 1, dot opacity for each cell, default 1}

\item{do.label}{a logical value indicating whether to add text to mark each cell label}

\item{label.size}{numeric, size of text labels, default 4}

\item{label.repel}{a logical value indicating whether to repel the labels to avoid overlapping, default False}

\item{label.plot}{cell labels to add text annotations, default NULL, add text for all labels}
}
\value{
a ggplot object
}
\description{
Produce a ggplot object with cells on 2D embedding, colored by given labels of each cell.
}
