% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCellScore}
\alias{plotCellScore}
\title{Plot a score for each cell}
\usage{
plotCellScore(X, score, cell.col = c("blue", "white", "red"),
  size = 0.5)
}
\arguments{
\item{X}{matrix, 2D embedding of each cell for the plot}

\item{score}{numeric vector, a single value to color each cell, continuous}

\item{cell.col}{string vector, color bar to use for "score", defaul c("blue","white","red")}

\item{size}{numeric, dot size for each cell, default 0.5}
}
\value{
a ggplot object
}
\description{
Produce a ggplot object with cells on 2D embedding, colored by a given score for each cell.
}
