% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDAcells.R
\name{updateDAcells}
\alias{updateDAcells}
\title{Update DA cells}
\usage{
updateDAcells(X, pred.thres = c(0.05, 0.95), do.plot = T,
  plot.embedding = NULL, size = 0.5)
}
\arguments{
\item{X}{output from getDAcells()}

\item{pred.thres}{length-2 vector, top and bottom threshold on the predictions from the
logistic classification, default c(0.05,0.95)}

\item{do.plot}{a logical value to indicate whether to return ggplot objects showing the results,
default True}

\item{plot.embedding}{size N-by-2 matrix, 2D embedding for the cells}

\item{size}{cell size to use in the plot, default 0.5}
}
\value{
a list of results with updated DA cells
}
\description{
Use different threshold to select DA cells based on an output from getDAcells()
}
